Added event and action processors as interfaces for ease of introducing own implementation
